==> datasets/1_ETL_Steam_Games.ipynb <==
# 🚀 ETL del dataset `output_steam_games`
#### En este notebook se desarrolla la 📦 extracción, 💱 transformación y 📥 carga del conjunto de datos `output_steam_games`

#### 📥Importaciones 

import pandas as pd
import json
import re 

#### 📦 Extracción de los datos y primera exploración

Se extraen los datos desde el archivo json, se convierte en Dataframe y se realiza una observación de su contenido.

# Ruta al dataset 
ruta_juegos = './output_steam_games.json'

# Se lee de cada línea del dataset
filas_juegos = []
with open(ruta_juegos, encoding='utf-8') as archivo:
    for line in archivo:
        data = json.loads(line)
        filas_juegos.append(data)

# Se convierte en dataframe
df_juegos = pd.DataFrame(filas_juegos)
df_juegos

#### 💱 Transformacion de los datos

Se observan las filas que contienen vacíos en todo el registro y se eliminan dichos registros.

df_juegos = df_juegos.dropna(how='all').reset_index(drop=True)
df_juegos.shape

Se revisan los tipos de datos por columna y la cantidad de nulos.

def verificar_tipo_datos(df):
    
    mi_dict = {"nombre_campo": [], "tipo_datos": [], "no_nulos_%": [], "nulos_%": [], "nulos": []}

    for columna in df.columns:
        porcentaje_no_nulos = (df[columna].count() / len(df)) * 100
        mi_dict["nombre_campo"].append(columna)
        mi_dict["tipo_datos"].append(df[columna].apply(type).unique())
        mi_dict["no_nulos_%"].append(round(porcentaje_no_nulos, 2))
        mi_dict["nulos_%"].append(round(100-porcentaje_no_nulos, 2))
        mi_dict["nulos"].append(df[columna].isnull().sum())

    df_info = pd.DataFrame(mi_dict)
            
    return df_info

verificar_tipo_datos(df_juegos)

Luego de borrar las filas NaN se observan 13 columnas y 32135 filas.

 Aún contienen nulos algunas columnas. 

 Se revisan algunas columnas para conocer su estructura.

# se observa el tipo de dato en 'genres'
df_juegos['genres'][0]

# se observa el tipo de dato en 'tags'
df_juegos['tags'][0]

# se observa el tipo de dato en 'specs'
df_juegos['specs'][100]

# se observa el tipo de dato en 'price'
df_juegos['price'].unique()

# se observa el tipo de dato en 'early_access'
df_juegos['early_access'].unique()

Las columnas que contiene este conjunto son:   

* **publisher**: es la empresa publicadora del contenido.
* **genres**: es el género del item, es decir, del juego. Esta formado por una lista de uno o mas géneros por registro.
* **app_name**: es el nombre del item, es decir, del juego.
* **title**: es el título del item.
* **url**: es la url del juego.
* **release_date**: es la fecha de lanzamiento del item en formato 0000-00-00.
* **tags**: es la etiqueta del contenido. Esta formado por una lista de uno o mas etiquetas por registro.
* **reviews_url**: es la url donde se encuentra el review de ese juego.
* **specs**: son especificaciones de cada item. Es una lista con uno o mas string con las especificaciones.
* **price**: es el precio del item.
* **early_access**: indica el acceso temprano con un True/False.
* **id**: es el identificador único del contenido.
* **developer**: es el desarrollador del contenido.

#### ✅​ Verificación de duplicados

Se analizan si hay duplicados teniendo en cuenta la columna del id del item.

def verifica_duplicados_por_columna(df, columna):
    # Se filtran las filas duplicadas
    duplicated_rows = df[df.duplicated(subset=columna, keep=False)]
    if duplicated_rows.empty:
        return "No hay duplicados"
    
    # se ordenan las filas duplicadas para comparar entre sí
    duplicated_rows_sorted = duplicated_rows.sort_values(by=columna)
    return duplicated_rows_sorted

verifica_duplicados_por_columna(df_juegos, 'id')

Se observa que solo hay dos items duplicados. 

Por un lado el 'id' 612880 que se observa es el mismo registro en los dos casos. Por lo que se puede eliminar cualquier de ellos. 

Por otro lado, se observa que la segunda dupla de duplicados el 'id' es NaN y la primer fila presenta mas cantidad de NaN que la segunda.



df_juegos[df_juegos['developer']=='Rocksteady Studios,Feral Interactive (Mac)']

Se puede observar que el item que no tiene 'id' es el 'id' 200260 y que el registro se encuentra completo. Por lo tanto, se pueden borrar ambas filas dupliacadas.

# se eliminan index duplicados
index_a_eliminar = [14573, 74, 30961]
df_juegos = df_juegos.drop(index_a_eliminar)

#### 💱 Transformación de la columna 'release_date'

Se observan las cantidades de registros por cada fecha

df_juegos['release_date'].value_counts()

Se crea una funcion para reemplazar valores los nulos por 'Dato no disponible'

def obtener_anio_release(fecha):
    if pd.notna(fecha):
        if re.match(r'^\d{4}-\d{2}-\d{2}$', fecha):
            return fecha.split('-')[0]
    return 'Dato no disponible'

Se crea una columna nueva con el año

df_juegos['release_anio'] = df_juegos['release_date'].apply(obtener_anio_release)

Se elimina la columna 'release_date'

df_juegos = df_juegos.drop('release_date', axis=1)
df_juegos.head()

df_juegos['release_anio'].unique()

#### 💱 Transformación de las columnas 'publisher', 'app_name', 'title', 'developer'

En estas columnas se observan valores nulos, por lo que se decide completarlos con valores del tipo 'Dato no disponible' para que el tipo de dato sea uno solo en la columna.

# Columnas a transformar
columnas_a_completar = ['publisher', 'app_name', 'title', 'developer']

# Se rellenan los nulos
df_relleno = df_juegos[columnas_a_completar].fillna('Dato no disponible')

# Se borran las columnas originales y se concatenan las rellenas con todo el dataframe
df_juegos = pd.concat([df_juegos.drop(columnas_a_completar, axis=1), df_relleno], axis=1)
df_juegos.head()

#### 💱 Transformación columna 'genres'

La columna 'genres' esta formada por una lista de los distintos géneros de los videojuegos. Se necesita hacer una columna con cada género.

df_juegos = df_juegos.explode('genres')
df_juegos = df_juegos.dropna(subset=['genres'])
df_juegos.head()

#### 💱 Transformación columnas 'tags', 'specs', 'url', 'reviews_url', 'price' y 'early_access'

Estas columnas no se utilizarán para las consultas de la API ni para el modelo de recomendación, por lo tanto se eliminan del dataframe.

df_juegos = df_juegos.drop(['tags', 'specs', 'url', 'reviews_url', 'price', 'early_access'], axis=1)
df_juegos.columns

#### ✅ Finalmente, se verifican el tipo de dato y si quedan nulos.

verificar_tipo_datos(df_juegos)

## 📥 Carga del dataset `output_steam_games.json`

Se guarda el dataframe transformado como `steam_games_limpio`.

archivo_limpio = 'Data/steam_games_limpio.csv'
df_juegos.to_csv(archivo_limpio, index=False, encoding='utf-8')
print(f'Se guardó el archivo {archivo_limpio}')
