==> datasets/2_ETL_Users_Items.ipynb <==
# 🚀 ETL del dataset `australian_users_items`
#### En este notebook se desarrolla la 📦 extracción, 💱 transformación y 📥 carga del conjunto de datos `australian_users_items`

#### 📥Importaciones 

import pandas as pd
import json
import ast 

#### 📦 Extracción de los datos y primera exploración

Se extraen los datos desde el archivo json, se convierte en Dataframe y se realiza una observación de su contenido.

# Ruta al dataset
ruta_items = './australian_users_items.json'

# Se lee de cada línea del dataset
filas_items = []
with open(ruta_items, encoding='utf-8') as archivo:
    for line in archivo.readlines():
        filas_items.append(ast.literal_eval(line))

# Se convierte en dataframe
df_items = pd.DataFrame(filas_items)
df_items

Se observan 5 columnas y 88309 filas. 

#### ✅ Se verifican los tipos de datos de cada columna y si hay nulos.

def verificar_tipo_datos(df):
    
    mi_dict = {"nombre_campo": [], "tipo_datos": [], "no_nulos_%": [], "nulos_%": [], "nulos": []}

    for columna in df.columns:
        porcentaje_no_nulos = (df[columna].count() / len(df)) * 100
        mi_dict["nombre_campo"].append(columna)
        mi_dict["tipo_datos"].append(df[columna].apply(type).unique())
        mi_dict["no_nulos_%"].append(round(porcentaje_no_nulos, 2))
        mi_dict["nulos_%"].append(round(100-porcentaje_no_nulos, 2))
        mi_dict["nulos"].append(df[columna].isnull().sum())

    df_info = pd.DataFrame(mi_dict)
            
    return df_info

verificar_tipo_datos(df_items)

La columna 'items' es una lista, así que se explora la misma para conocer su estructura.

df_items['items'][0]

Las columnas que contiene este conjunto son:

* **user_id**: es un identificador único del usuario.
* **items_count**: es un número entero que indica la cantidad de juegos que consumio el usuario.
* **steam_id**: es un número único para la plataforma.
* **user_url**: es la url del perfil del usuario
* **items**: es una lista de uno o mas diccionarios de los items que consume cada usuario. Cada diccionario tiene las siguientes claves:
  * **item_id**: es el identificador del item, es decir, del juego.
  * **item_name**: es el nombre del contenido que consume, es decir, del juego.
  * **playtime_forever**: es el tiempo acumulado que un usuario jugó a un juego.
  * **playtime_2weeks**: es el tiempo acumulado que un usuario jugó a un juego en las últimas dos semanas.

#### 💱 Transformación de la columna 'items'

La columna 'items' se encuentra anidada, es decir, es una lista de diccionarios. 

Se normaliza la columna 'items' para obtener una columna por cada clave de ese diccionario, manteniendo la trazabilidad de las columnas 'steam_id','items_count','user_id' y 'user_url'.

df_items2 = pd.json_normalize(filas_items, record_path=['items'], meta=['steam_id','items_count','user_id', 'user_url'] )
df_items2

#### ✅ Se verifican los tipos de datos y si hay nulos.

verificar_tipo_datos(df_items2)

Se observa si hay duplicados.

duplicados = df_items2.loc[df_items2.duplicated()]
duplicados

Se observan 59104 duplicados, por lo que se decide borrarlos.

df_items2 = df_items2.drop_duplicates(keep='first')
df_items2

Se elimina la columna 'playtime_2weeks' porque la información es sobre el tiempo de juego acumulado por cada usuario, que se encuentra en 'playtime_forever'.

df_items2 = df_items2.drop('playtime_2weeks', axis=1)
df_items2.columns

## 📥 Carga del dataset `australian_users_items.json`

Se guarda el dataframe transformado como `users_items_limpio.csv`.

archivo_limpio = 'Data/users_items_limpio.csv'
df_items2.to_csv(archivo_limpio, index=False, encoding='utf-8')
print(f'Se guardó el archivo {archivo_limpio}')
