==> datasets/3_ETL_User_Reviews.ipynb <==
# 🚀 ETL del dataset `australian_user_reviews`
#### En este notebook se desarrolla la 📦 extracción, 💱 transformación y 📥 carga del conjunto de datos `australian_user_reviews`

#### 📥Importaciones 

import pandas as pd
import json
import ast
import re

#### 📦 Extracción de los datos y primera exploración

Se extraen los datos desde el archivo json, se convierte en Dataframe y se realiza una observación de su contenido.

# Ruta al dataset 
ruta_review = './australian_user_reviews.json'

# Se lee de cada línea del dataset
filas_review = []
with open(ruta_review, encoding='utf-8') as archivo:
    for line in archivo.readlines():
        filas_review.append(ast.literal_eval(line))

# Se convierte en dataframe
df_reviews = pd.DataFrame(filas_review)
df_reviews

Se observan 3 columnas y 25799 filas. 

#### ✅ Se verifican los tipos de datos de cada columna y si hay nulos.

def verificar_tipo_datos(df):
    
    mi_dict = {"nombre_campo": [], "tipo_datos": [], "no_nulos_%": [], "nulos_%": [], "nulos": []}

    for columna in df.columns:
        porcentaje_no_nulos = (df[columna].count() / len(df)) * 100
        mi_dict["nombre_campo"].append(columna)
        mi_dict["tipo_datos"].append(df[columna].apply(type).unique())
        mi_dict["no_nulos_%"].append(round(porcentaje_no_nulos, 2))
        mi_dict["nulos_%"].append(round(100-porcentaje_no_nulos, 2))
        mi_dict["nulos"].append(df[columna].isnull().sum())

    df_info = pd.DataFrame(mi_dict)
            
    return df_info

verificar_tipo_datos(df_reviews)

Se observa si hay o no duplicados por cada 'user_id'.

def verifica_duplicados_por_columna(df, columna):

    # Se filtran las filas duplicadas
    duplicated_rows = df[df.duplicated(subset=columna, keep=False)]
    if duplicated_rows.empty:
        return "No hay duplicados"
    
    # se ordenan las filas duplicadas para comparar entre sí
    duplicated_rows_sorted = duplicated_rows.sort_values(by=columna)
    return duplicated_rows_sorted

filas_duplicadas = verifica_duplicados_por_columna(df_reviews, 'user_id')
filas_duplicadas

Se observan 623 filas duplicadas en la columna 'user_id'

Se revisan si los review dentro de los datos anidados de 'review' se encuentran duplicados o si solo se duplica el 'user_id' porque hay mas de un comentario realizado por ese usuario.

user_id = '05041129'
user_reviews = filas_duplicadas[filas_duplicadas['user_id'] == user_id]['reviews']

for review_list in user_reviews:
    for review in review_list:
        print(review['review'])
    print('-' * 100)

#### 💱 Transformacion de los datos

Se puede ver que los review son los mismos para cada registro, por lo que se borran los duplicados, dejando la primer ocurrencia de los registros.

df_reviews = df_reviews.drop_duplicates(subset='user_id', keep='first')
verifica_duplicados_por_columna(df_reviews, 'user_id')

Se revisa la columna 'review' para entender el tipo de dato.

df_reviews['reviews'][0]

Las columnas de este conjunto son:

* **user_id**: es un identificador único para el usuario.
* **user_url**: es la url del perfil del usuario en streamcommunity.
* **reviews**: contiene una lista de diccionarios. Para cada usuario se tiene uno o mas diccionario con el review. Cada diccionario contiene:
    * **funny**: indica si alguien puso un emoticón de gracioso al review.
    * **posted**: es la fecha de posteo del review en formato Posted Month 00, 0000.
    * **last_edited**: es la fecha de la última edición.
    * **item_id**: es el identificador único del item, es decir, del juego.
    * **helpful**: es la estadística donde otros usuarios indican si fue útil la información.
    * **recommend**: es un booleano que indica si el usuario recomienda o no el juego.
    * **review**: es una sentencia string con los comentarios sobre el juego.

#### 💱 Transformación de la columna 'reviews'

La columna 'reviews' se presenta anidada, siendo una lista con uno o mas diccionarios como elementos. 

Se genera una columna por cada diccionario para posteriormente hacer un registro.

# Se transforma a columnas cada elemento de las listas
df_reviews2 = pd.json_normalize(df_reviews['reviews'])
df_reviews2.head()

En esta transformación, se pierde el 'user_id' y 'user_url' al que pertenece cada diccionario, pero aún mantiene la misma posición. Por lo que se concatena con el dataframe anterior, para no perder estos datos.

# Se agrega el 'user_id' y 'user_url' a las columnas separadas 
df_reviews2 = pd.concat([df_reviews[['user_id', 'user_url']], df_reviews2], axis=1)
df_reviews2.head()

Ahora se tienen los diccionarios por columnas, con el usuario que genera dicha información.

Se genera un registro por cada diccionario, manteniendo en cada caso el usuario que lo genera.

# Se utiliza pd.melt para transformar las columnas en filas conservando el 'user_id' y 'user_url'
df_reviews2 = pd.melt(df_reviews2, id_vars=['user_id', 'user_url'], 
                       value_vars=list(range(9)),
                       value_name='reviews')
df_reviews2.head()

Se observa que quedan registros None. Esto ocurre porque hay usuarios que hicieron mas reviews que otros. En este ejemplo se puede ver este caso:

df_reviews2[df_reviews2['user_id']=='76561197970982479']

Se eliminan los registros que tienen None en 'reviews'.

df_reviews2 = df_reviews2.dropna()

Se verifica que solo queden el 'user_id' con la cantidad de diccionarios que le corresponden

df_reviews2[df_reviews2['user_id']=='76561197970982479']

Se convierte cada diccionario en columna.

# Se separan por columnas cada una de las claves de 'reviews'
df_reviews = df_reviews2['reviews'].apply(pd.Series, dtype='object')
df_reviews = df_reviews.add_prefix('reviews_')
df_reviews.head()

Se puede observar que la columna de 'user_id' y 'user_url' se perdieron nuevamente, por lo que se vuelve a concatenar.

# Se une con el 'user_id' y 'user_url'
df_reviews = pd.concat([df_reviews2[['user_id', 'user_url']], df_reviews], axis=1)
df_reviews.head()

Se observa que hay valores faltantes en algunas columnas, pero no como nulos, sino que tienen un espacio:

df_reviews['reviews_last_edited'][0]

Se reemplazan esos espacios como valores nulos.

df_reviews.replace('', None, inplace=True)
df_reviews.head()

Se analizan los tipos de datos y los nulos que quedaron después de desanidar la columna 'reviews'.

verificar_tipo_datos(df_reviews)

Se observa entre un 86% a 89% de nulos en las columnas 'reviews_funny' y 'reviews_last_edited' por lo que se eliminan estas columnas. 

Por lado se observa un 0,05% de nulos en la columna 'reviews_review', pero no se eliminan esos registros ya que se consideran como un comentario neutral.

# Se eliminan las columnas 'reviews_funny' y 'reviews_last_edited'
df_reviews = df_reviews.drop(columns=['reviews_funny', 'reviews_last_edited'])
df_reviews.columns

#### 💱 Transformación de la columna 'reviews_posted'

La fecha donde se hizo el posteo de la review se encuentra en formato: `Posted November 5, 2011.` , se transforma la fecha a este formato: `YYYY-MM-DD`. 

def convertir_fecha(cadena_fecha):
    match = re.search(r'(\w+\s\d{1,2},\s\d{4})', cadena_fecha)
    if match:
        fecha_str = match.group(1)
        try:
            fecha_dt = pd.to_datetime(fecha_str)
            return fecha_dt.strftime('%Y-%m-%d')
        except:
            return 'Fecha inválida'
    else:
        return 'Formato inválido'

df_reviews['reviews_date'] = df_reviews['reviews_posted'].apply(convertir_fecha)
df_reviews['reviews_date']

Se observa que hay registros que contienen un formato inválido distinto a los demas. Estos registros no se podrán consultar desde la API, pero las demás columnas serán útiles para aportar información.

df_reviews[df_reviews['reviews_date'] == 'Formato inválido']

Se elimina la columna 'reviews_posted' porque solo indica el día y el mes del posteo.

df_reviews = df_reviews.drop('reviews_posted', axis=1)
df_reviews.columns

#### 💱 Transformación de la columna 'reviews_review'

Esta columna tiene un 5% de valores nulos, por lo que se eliminan.

df_reviews = df_reviews.dropna(subset=['reviews_review'])
verificar_tipo_datos(df_reviews)

## 📥 Carga del dataset `australian_user_reviews`

Se guarda el conjunto de datos transformado como `user_review_limpio`. 

archivo_limpio = 'Data/user_review_limpio.csv'
df_reviews.to_csv(archivo_limpio, index=False, encoding='utf-8')
print(f'Se guardó el archivo {archivo_limpio}')
