==> datasets/4_Feature_Engineering.ipynb <==
# 🚀​ Preparación de conjuntos de datos para consultas
En este notebook se preparan datos necesarios para hacer las consultas en la API.

#### 📥Importaciones 

from textblob import TextBlob
import pandas as pd
import json
import ast 
import re
import pyarrow as pa
import pyarrow.parquet as pq 

#### 📦 Extracción de los datos

Se extrean los datos de `df_juegos` que contiene la información relacionada a los juegos que tiene la plataforma.

df_juegos = pd.read_csv('Data/steam_games_limpio.csv', encoding='utf-8')

Se extrean los datos de `df_items` que contiene la información relacionada al consumo de los juegos por parte de los usuarios.

df_items = pd.read_csv('Data/users_items_limpio.csv', encoding='utf-8')

Se extrean los datos de `df_reviews` que contiene la información relacionada a los usuarios que realizaron review de los juegos que consumen.

df_reviews = pd.read_csv('Data/user_review_limpio.csv', encoding='utf-8')

### 👍​​👎​ Análisis de sentimientos

Se pide crear una nueva columna llamada 'sentiment_analysis' que reemplace a 'reviews_review' donde se realice un análisis de sentimiento de los comentarios con la siguiente escala:

* 0 si es malo.
* 1 si es neutral.
* 2 si es positivo.

Dado que el objetivo de este proyecto es realizar una prueba de concepto, consiguiendo un producto mínimo viable, se realiza un análisis de sentimiento básico utilizando ``TextBlob`` que es una biblioteca de procesamiento de lenguaje natural (NLP) en Python. El objetivo de esta metodología es asignar un valor numérico a un texto, en este caso a los comentarios que los usuarios dejaron para un juego determinado, para representar si el sentimiento expresado en el texto es negativo, neutral o positivo. 

Esta metodología toma una revisión de texto como entrada, utiliza ``TextBlob`` para calcular la polaridad de sentimiento y luego clasifica la revisión como negativa, neutral o positiva en función de la polaridad calculada. En este caso, se consideraron las polaridades por defecto del modelo, el cuál utiliza umbrales -0.2 y 0.2, siendo polaridades negativas por debajo de -0.2, positivas por encima de 0.2 y neutrales entre medio de ambos.

Primero se crea una funcion para que realice el análisis de sentimiento en un texto dado y devuelva un valor numérico que represente el sentimiento.

def analisis_sentimiento(review):
    if review is None:
        return 1
    analysis = TextBlob(review)
    polarity = analysis.sentiment.polarity
    if polarity < -0.2:
        return 0  
    elif polarity > 0.2: 
        return 2 
    else:
        return 1 

df_reviews['sentiment_analysis'] = df_reviews['reviews_review'].apply(analisis_sentimiento)
df_reviews.head()

Se crea una función que imprime ejemplos de reviews para cada categoría de análisis de sentimiento.

def ejemplos_review_por_sentimiento(reviews, sentiments):
    for sentiment_value in range(3):
        print(f"Para la categoría de análisis de sentimiento {sentiment_value} se tienen estos ejemplos de reviews:")
        sentiment_reviews = [review for review, sentiment in zip(reviews, sentiments) if sentiment == sentiment_value]
        
        for i, review in enumerate(sentiment_reviews[:3], start=1):
            print(f"Review {i}: {review}")
        
        print("\n")

Se revisan algunos ejemplos para cada una de las clases de sentimiento.

ejemplos_review_por_sentimiento(df_reviews['reviews_review'], df_reviews['sentiment_analysis'])

Se elimina la columna de 'reviews_review'.

df_reviews = df_reviews.drop(columns=['reviews_review'])
df_reviews.columns

Se revisan los tipos de datos por columna y la cantidad de nulos.

def verificar_tipo_datos(df):
    
    mi_dict = {"nombre_campo": [], "tipo_datos": [], "no_nulos_%": [], "nulos_%": [], "nulos": []}

    for columna in df.columns:
        porcentaje_no_nulos = (df[columna].count() / len(df)) * 100
        mi_dict["nombre_campo"].append(columna)
        mi_dict["tipo_datos"].append(df[columna].apply(type).unique())
        mi_dict["no_nulos_%"].append(round(porcentaje_no_nulos, 2))
        mi_dict["nulos_%"].append(round(100-porcentaje_no_nulos, 2))
        mi_dict["nulos"].append(df[columna].isnull().sum())

    df_info = pd.DataFrame(mi_dict)
            
    return df_info

verificar_tipo_datos(df_reviews)

#### 💱 Transformacion de `df_reviews`: agregamos la columna 'release_anio'

Se agrega el titulo, el año de lanzamiento de un juego y la empresa desarrolldora del mismo al dataframe `df_reviews`. En primer lugar se extraen las columnas de id del juego, el titulo del juego y el año de lanzamiento del mismo, se borran los duplicados y luego se unen con el dataframe de reviews.

anio_lanzamiento_item = df_juegos[['id', 'title', 'release_anio', 'developer']]
# Se renombra la columna 'id' para unirla con el dataframe anterior
anio_lanzamiento_item = anio_lanzamiento_item.rename(columns={'id':'reviews_item_id'})
# se eliminan los duplicados
anio_lanzamiento_item = anio_lanzamiento_item.drop_duplicates()
anio_lanzamiento_item

df_reviews = df_reviews.merge(anio_lanzamiento_item, on='reviews_item_id')
df_reviews

Se elimina la columna 'user_url'.

df_reviews.drop('user_url', axis=1, inplace=True)
df_reviews

## 📥 Carga de los dataframe

dfs = [df_reviews, df_juegos, df_items]
# Nombres correspondientes a cada DataFrame
names = ['df_reviews', 'df_games', 'df_items']

for df, name in zip(dfs, names):
    archivo = f'data/{name}_unido.csv'
    df.to_csv(archivo, index=False, encoding='utf-8')
    print(f"DataFrame '{name}' guardado como '{archivo}'")

Para optimizar la estructura de los datos en el deploy, se aprovecha en este punto a guardar los dataframe en formato parquet.

for df, name in zip(dfs, names):
    archivo = f'data/parquet/{name}.parquet'
    pq.write_table(pa.Table.from_pandas(df), archivo)
    print(f"DataFrame '{name}' guardado como '{archivo}'")
