==> datasets/7_Sistema_de_recomendacion.ipynb <==
# 🚀 Sistema de recomendación de juegos

En este notebook se hace la experimentación para encontrar dos modelos de recomendación, que generen una lista de 5 juegos ya sea ingresando el nombre de un juego o el id de un usuario.

En el primer caso, el modelo tiene una relación ítem-ítem, esto es se toma un item, en base a que tan similar es ese ítem al resto, se recomiendan similares. En el segundo caso, el sistema de recomendación aplica el filtro user-item, esto es tomar un usuario, se encuentran usuarios similares y se recomiendan ítems que a esos usuarios similares les gustaron.

#### 📥Importaciones 

import pandas as pd
import numpy as np
import scipy as sp
import operator
from sklearn.metrics.pairwise import cosine_similarity
import pyarrow as pa
import pyarrow.parquet as pq

#### 📦 Se extraen los datos que se prepararon luego del EDA y se convierten en dataframe para ser utilizados por el modelo.

df = pd.read_csv('data/df_recomendacion.csv')
df

Se un dataframe que contiene los 'user_id' como indices, los juegos ('item_name') como columnas y los ´rating' como valores.

piv = df.pivot_table(index=['user_id'], columns=['item_name'], values='rating')
piv

Se normalizan los valores del dataframe `piv` para restar la media de las calificaciones de un usuario y despues dividir por la diferencia entre el valor máximo y mínimo de las calificaciones. Esto ajusta las calificaciones de un usuario de manera que estén centradas en cero y escaladas en función de su variabilidad. Los usuarios que solo dieron una calificación o calificaron todos los juegos de la misma manera van a ser eliminados durante este proceso de normalización. Esto se debe a que estos usuarios no aportan información útil para el modelo de recomendación si todas sus calificaciones son iguales o si solo tienen una calificación.

# Normalización del dataframe 'piv'
piv_norm = piv.apply(lambda x: (x-np.mean(x))/(np.max(x)-np.min(x)), axis=1)
# Se borran las columnas que contienen solo cero o no tienen rating, se rellenan los vacios con 0 y se hace la transpuesta
piv_norm.fillna(0, inplace=True)
piv_norm = piv_norm.T
piv_norm = piv_norm.loc[:, (piv_norm != 0).any(axis=0)]
piv_norm

A los datos de esta matriz normalizada se los convierte a un formato de matriz dispersa (sparse matrix) para reducir la memoria utilizada y mejorar la eficiencia en el manejo de grandes conjuntos de datos, especialmente cuando la mayoría de los valores en la matriz son ceros. 

piv_sparse = sp.sparse.csr_matrix(piv_norm.values)
piv_sparse

Se crean dos matrices de similitud utilizando la similitud del coseno para medir la similitud entre los juegos (item_similarity) y entre los usuarios (user_similarity).

La similitud del coseno es una medida comúnmente utilizada para evaluar la similitud entre dos vectores en un espacio multidimensional. En el contexto de sistemas de recomendación y análisis de datos, la similitud del coseno se utiliza para determinar cuán similares son dos conjuntos de datos o elementos, y se calcula utilizando el coseno del ángulo entre los vectores que representan esos datos o elementos.

item_similarity = cosine_similarity(piv_sparse)
user_similarity = cosine_similarity(piv_sparse.T)

Se insertan las matrices anteriores en un Dataframe para estructurar y organizar los resultados de manera más accesible y comprensible.

#item similarity dataframe
item_sim_df = pd.DataFrame(item_similarity, index = piv_norm.index, columns = piv_norm.index)
#user similarity dataframe
user_sim_df = pd.DataFrame(user_similarity, index = piv_norm.columns, columns = piv_norm.columns)

## 🚀 Función de recomendación según un juego

Se realiza una función que devuelve una recomendación de 5 juegos en función de un juego dado, teniendo en cuenta los valores más altos de similitud del coseno. 

def recomendacion_juego(game):
    count = 1
    print('Similar games to {} include:\n'.format(game))
    for item in item_sim_df.sort_values(by = game, ascending = False).index[1:6]:
        print('No. {}: {}'.format(count, item))
        count +=1  

Se prueba la función 'recomendacion_juego'

recomendacion_juego('Killing Floor')

recomendacion_juego('theHunter')

## 🚀 Función de recomendación según un usuario

Se realiza una función que genera una lista de 5 juegos recomendados para un usuario en función de las calificaciones de usuarios similares. Los juegos que son más frecuentemente recomendados por usuarios similares se consideran como las principales recomendaciones para ese usuario.

def recomendacion_usuario(user):
    
    # Se verifica si el usuario está presente en las columnas de piv_norm (si no está, devuelve un mensaje)
    if user not in piv_norm.columns:
        return('No data available on user {}'.format(user))
    
    # Se obtienen los usuarios más similares al usuario dado
    sim_users = user_sim_df.sort_values(by=user, ascending=False).index[1:11]
    
    best = []  # Lista para almacenar los juegos mejor calificados por usuarios similares
    most_common = {}  # Diccionario para contar cuántas veces se recomienda cada juego
    
    # Para cada usuario similar, se encuentra el juego mejor calificado y se agrega a la lista 'best'
    for i in sim_users:
        max_score = piv_norm.loc[:, i].max()
        best.append(piv_norm[piv_norm.loc[:, i]==max_score].index.tolist())
    
    # Se cuenta cuántas veces se recomienda cada juego
    for i in range(len(best)):
        for j in best[i]:
            if j in most_common:
                most_common[j] += 1
            else:
                most_common[j] = 1
    
    # Se ordenan los juegos por la frecuencia de recomendación en orden descendente
    sorted_list = sorted(most_common.items(), key=operator.itemgetter(1), reverse=True)
    
    # Se devuelven los 5 juegos más recomendados
    return sorted_list[:5]  

# ejemplo
recomendacion_usuario('zvanik')

# ejemplo
recomendacion_usuario('76561197970982479')

## 📥 Carga de las matrices para recomendar en la API

Se guardan en formato parquet lo que permite una compresión y codificación eficiente.

pq.write_table(pa.Table.from_pandas(piv_norm), 'data/parquet/piv_norm.parquet')
pq.write_table(pa.Table.from_pandas(user_sim_df), 'data/parquet/user_sim_df.parquet')
pq.write_table(pa.Table.from_pandas(item_sim_df), 'data/parquet/item_sim_df.parquet')
print('Se guardaron correctamente')
